requirements.txt
google-generativeai>=0.4.0
pydantic>=2.5.3
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
